+++ packages/core/src/server.ts
@@ -59,6 +59,7 @@ import { PBES2Container } from "./container";
 import { KeyStoreEntry } from "./key-store";
 import { Config, ConfigParam } from "./config";
 import { Provisioner, Provisioning, ProvisioningStatus, StubProvisioner } from "./provisioning";
+import { DefaultScimProvider } from "./scim";
 import { V3Compat } from "./v3-compat";
 
 /** Server configuration */
@@ -152,6 +153,10 @@ export class Controller extends API {
 
     get provisioner() {
         return this.server.provisioner;
+    }
+
+    get scim() {
+        return this.server.scim;
     }
 
     async authenticate(req: Request, ctx: Context) {
@@ -1865,6 +1870,7 @@ export class Server {
         /** Attachment storage */
         public attachmentStorage: AttachmentStorage,
         public provisioner: Provisioner = new StubProvisioner(),
+        public scim: DefaultScimProvider,
         public legacyServer?: LegacyServer
     ) {}
