+++ packages/pwa/webpack.config.js
@@ -30,9 +30,6 @@ module.exports = {
         alias: {
             assets: assetsDir,
         },
-        fallback: {
-            http: false,
-        },
     },
     module: {
         rules: [
